backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.3
django-allauth==0.61.1
django-cors-headers==4.3.1
django-oauth-toolkit==2.3.0
django-rest-framework-social-oauth2==1.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
google-auth==2.28.2
gunicorn==21.2.0
h11==0.14.0
idna==3.6
jwcrypto==1.5.6
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==1.7.1
PySocks==1.7.1
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
social-auth-app-django==4.0.0
social-auth-core==4.0.2
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
webdriver-manager==4.0.1
whitenoise==6.6.0
wsproto==1.2.0
